Suppression of warnings.

+++ src/INTERP_KERNEL/DirectedBoundingBox.cxx
@@ -20,7 +20,6 @@
 // Created   : Mon Apr 12 14:41:22 2010
 // Author    : Edward AGAPOV (eap)
 
-
 #include "DirectedBoundingBox.hxx"
 
 #include "InterpolationUtils.hxx"
@@ -237,9 +236,9 @@ namespace INTERP_KERNEL
     if ( dim > 1 )
       {
         for ( coord = pts; coord < coordEnd; )
-          for ( int i = 0; i < dim; ++i )
+          for ( int i = 0; i < (int)dim; ++i )
             gc[i] += *coord++;
-        for ( int j = 0; j < dim; ++j )
+        for ( int j = 0; j < (int)dim; ++j )
           gc[j] /= numPts;
 
       }
@@ -309,9 +308,9 @@ namespace INTERP_KERNEL
     if ( dim > 1 )
       {
         for ( unsigned i = 0; i < numPts; ++i )
-          for ( int j = 0; j < dim; ++j )
+          for ( int j = 0; j < (int)dim; ++j )
             gc[j] += pts[i][j];
-        for ( int j = 0; j < dim; ++j )
+        for ( int j = 0; j < (int)dim; ++j )
           gc[j] /= numPts;
       }
 
@@ -569,7 +568,7 @@ namespace INTERP_KERNEL
                                        const double*        minmax) const
   {
     int iC, nbCorners = 1;
-    for ( int i=0;i<_dim;++i ) nbCorners *= 2;
+    for ( int i=0;i<(int)_dim;++i ) nbCorners *= 2;
     corners.resize( nbCorners * _dim );
     // each coordinate is filled with either min or max, nbSwap is number of corners
     // after which min and max swap
@@ -662,7 +661,7 @@ namespace INTERP_KERNEL
       for ( int iC = 0, nC = cornersThis.size()/_dim; iC < nC; ++iC)
         {
           fromLocalCS( &cornersThis[iC*_dim], globCorner );
-          for ( int i = 0; i < _dim; ++i )
+          for ( int i = 0; i < (int)_dim; ++i )
             {
               if ( globCorner[i] < mmBox._minmax[i*2] )   mmBox._minmax[i*2] = globCorner[i];
               if ( globCorner[i] > mmBox._minmax[i*2+1] ) mmBox._minmax[i*2+1] = globCorner[i];
@@ -715,7 +714,7 @@ namespace INTERP_KERNEL
       data.insert( data.end(), &_axes[0], &_axes[0] + _axes.size());
       data.insert( data.end(), &_minmax[0], &_minmax[0] + _minmax.size());
     }
-    if ( data.size() < dataSize( _dim ))
+    if ( data.size() < (unsigned)dataSize( _dim ))
       data.resize( dataSize( _dim ), 0 );
     return data;
   }
